CARP Utils: fix precision bug introduced by testcase PivotCalc11.
In PivotCalc11, OOB items at indices 300 and 301 are extremely close. This leads to pivots 2406 and 2407 being identical if floats are used, as subdividing a range that small is beyond the scope of floats. We use doubles wherever pivot calculation is involved to fix this.

+++ src/carp/carp_utils.h
@@ -92,7 +92,7 @@ class PivotUtils {
   static int GetRangeBounds(PivotCalcCtx* pvt_ctx, float& range_start,
                             float& range_end);
 
-  static float WeightedAverage(float a, float b, float frac);
+  static double WeightedAverage(double a, double b, double frac);
 
   friend class RangeUtilsTest;
 };
